requirements.txt
python-dotenv
pandas
numpy
setuptools
# pymongo[srv]==3.6
pymongo>=4.0
certifi